fix manage page input close modal problem, and extract modal component to another file
