+++ src/Pages/ManagePage/Manage.js
@@ -17,6 +17,7 @@ import Swal from "sweetalert2";
 import items from "../../images/items.jpg";
 import Popup from "reactjs-popup";
 import "reactjs-popup/dist/index.css";
+import EditShareList from "./EditShareList.js";
 
 // const StyledModal = Modal.styled`
 //   height: 20rem;
@@ -30,21 +31,6 @@ import "reactjs-popup/dist/index.css";
 //   position: relative;
 //   /*transition: all 0.3s ease-in-out;*/
 // `;
-
-const StyledPopup = styled(Popup)`
-  &-content {
-    height: 20rem;
-    width: 20rem;
-    align-items: center;
-    justify-content: center;
-    background-color: white;
-    /* border: none; */
-    border-radius: 10px;
-    /*opacity: ${(props) => props.opacity};*/
-    position: relative;
-    /*transition: all 0.3s ease-in-out;*/
-  }
-`;
 
 // const customStyles = {
 //   content: {
@@ -256,103 +242,6 @@ const HistoryTrips = styled.div`
   color: #eedd42;
   /* background-color: rgb(57, 80, 73, 0.3); */
 `;
-const OpenButton = styled.button`
-  cursor: pointer;
-  font-family: "QuickSand";
-  font-size: 16px;
-  margin: 5px;
-  border: none;
-  border-radius: 40px;
-  height: 40px;
-  background-color: rgb(255, 255, 255, 0.3);
-  border: 1px solid lightgrey;
-  :hover {
-    color: white;
-    background-color: #eedd42;
-    border: 1px solid #eedd42;
-  }
-`;
-const AddByEmail = styled.div`
-  display: flex;
-  margin-top: 20px;
-  margin-left: 20px;
-`;
-const Form = styled.form``;
-const Input = styled.input`
-  font-family: "QuickSand";
-  width: 220px;
-  margin-right: 10px;
-  height: 25px;
-  font-size: 14px;
-  color: grey;
-  padding-left: 12px;
-  outline: none;
-  border-radius: 15px;
-  border: 1px solid #91ccb9;
-`;
-const Add = styled.button`
-  font-family: "QuickSand";
-  border-radius: 30px;
-  border: 1px solid #91ccb9;
-  background-color: white;
-  color: #91ccb9;
-  :hover {
-    background-color: #91ccb9;
-    border: 1px solid #91ccb9;
-    color: white;
-    cursor: pointer;
-  }
-`;
-const AlreadySharedList = styled.div`
-  width: 90%;
-  margin: 0 auto;
-`;
-const EachUser = styled.div`
-  display: flex;
-  margin-top: 5px;
-  justify-content: space-between;
-`;
-const Title = styled.div`
-  margin: 0 auto;
-  text-align: center;
-  margin-top: 20px;
-  border-bottom: 1px solid grey;
-`;
-const CloseModalBtn = styled.div``;
-const AllUsers = styled.div``;
-const TypeEmail = styled.div``;
-const Delete = styled.button`
-  font-family: "QuickSand";
-  border-radius: 20px;
-  border: 1px solid #91ccb9;
-  background-color: white;
-  color: #91ccb9;
-  height: 25px;
-  :hover {
-    cursor: pointer;
-    background-color: #91ccb9;
-    color: white;
-  }
-`;
-const CloseBtn = styled.button`
-  font-family: "QuickSand";
-  font-size: 16px;
-  border-radius: 20px;
-  height: 30px;
-  border: 1px solid #eedd42;
-  background-color: white;
-  color: grey;
-  position: absolute;
-  bottom: 20px;
-  left: 50%;
-  transform: translate(-50%);
-  :hover {
-    background-color: #eedd42;
-    border: 1px solid #eedd42;
-    color: black;
-    cursor: pointer;
-  }
-`;
 
 // const FadingBackground = styled(BaseModalBackground)`
 //   opacity: ${(props) => props.opacity};
@@ -371,130 +260,9 @@ function ManageSchedule() {
   const [tripEdit, setTripEdit] = useState([]);
   const [profileData, setProfileData] = useState([]);
   const [ownerEmail, setOwnerEmail] = useState([]);
-  const [shareEmail, setShareEmail] = useState([]);
 
   if (UID === undefined) {
     history.push(`/error`);
-  }
-
-  function EditShareList({ tripID, share }) {
-    // const [isOpen, setIsOpen] = useState(false);
-    const [open, setOpen] = useState(false);
-    const closeModal = () => setOpen(false);
-    //(props)
-    // let {tripID, share} = props;
-
-    // function toggleModal(e) {
-    //   setOpacity(0);
-    //   setIsOpen(!isOpen);
-    // }
-
-    // function afterOpen() {
-    //   setTimeout(() => {
-    //     setOpacity(1);
-    //   }, 100);
-    //   setCurrentShareEmail(share);
-    // }
-
-    // function beforeClose() {
-    //   return new Promise((resolve) => {
-    //     setOpacity(0);
-    //     setTimeout(resolve, 300);
-    //   });
-    // }
-
-    console.log(tripID, share);
-
-    return (
-      <div>
-        {/* {share
-          ? share.map((each, i) => {
-              console.log(each);
-              return <div>{each}</div>;
-            })
-          : null} */}
-
-        <StyledPopup
-          key={tripID}
-          open={open}
-          closeOnDocumentClick
-          onClose={closeModal}
-          // afterOpen={afterOpen}
-          // beforeClose={beforeClose}
-          // onBackgroundClick={toggleModal}
-          // onEscapeKeydown={toggleModal}
-          // opacity={opacity}
-          // backgroundProps={{ opacity }}
-          trigger={
-            <OpenButton
-              onClick={() => {
-                // setIsOpen(true);
-                setOpen(true);
-              }}
-            >
-              Edit
-            </OpenButton>
-          }
-          modal
-          position="right center"
-        >
-          {(close) => (
-            <>
-              <AddByEmail>
-                <Form>
-                  <Input
-                    type="text"
-                    name="email"
-                    id="inputEmail"
-                    placeholder="email"
-                    // onSummit={(e) => setShareEmail(e.target.value)}
-                  ></Input>
-                </Form>
-                <Add onClick={() => addUser(tripID, shareEmail)}>add</Add>
-              </AddByEmail>
-
-              <AlreadySharedList>
-                <Title>Share list</Title>
-                <CloseModalBtn></CloseModalBtn>
-                <AllUsers>
-                  {share
-                    ? share.map((each, i) => {
-                        /* console.log(each); */
-
-                        return (
-                          <EachUser key={i}>
-                            <TypeEmail>{each}</TypeEmail>
-                            <Delete
-                              onClick={() => deleteUser(tripID, shareEmail)}
-                            >
-                              delete
-                            </Delete>
-                          </EachUser>
-                        );
-                      })
-                    : null}
-                </AllUsers>
-              </AlreadySharedList>
-              <CloseBtn
-                // onClick={closeModal}
-                onClick={close}
-              >
-                Finish
-              </CloseBtn>
-            </>
-          )}
-        </StyledPopup>
-      </div>
-    );
-  }
-
-  function addUser(tripID, shareEmail) {
-    console.log(tripID, shareEmail);
-
-    // addShareEmail(tripID, email);
-  }
-  function deleteUser(tripID, email) {
-    removeShareEmail(tripID, email);
   }
 
   useEffect(() => {
@@ -564,14 +332,14 @@ function ManageSchedule() {
   //     console.log(doc);
   //   });
 
-  const renderEventsData = async () => {
+  const getEventsData = async () => {
     let tripDataEdit = await getTripDataByCanEdit(UID); //profileEmail //UID
     console.log(tripDataEdit);
     setTripEdit(tripDataEdit);
   };
 
   useEffect(() => {
-    renderEventsData();
+    getEventsData();
 
     //eslint-disable-next-line
   }, []);
@@ -665,10 +433,6 @@ function ManageSchedule() {
                   const time = trip[1].createTime;
                   const share = trip[1].share;
                   const tripID = trip[0];
-
-                  /* console.log(share); */
-
-                  /* const owner = trip[1].owner; */
 
                   return (
                     <Tr key={i}>
