+++ src/index.css
@@ -1,6 +1,5 @@
 body {
   margin: 0;
-  font-family: "Quicksand", sans-serif, "Caveat", cursive, "Allura", cursive;
   font-family: "Quicksand", sans-serif, "Caveat", cursive, "Allura", cursive,
     /* -apple-system, BlinkMacSystemFont,  */ "Segoe UI", "Roboto", "Oxygen",
     "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
